requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
MouseInfo==0.1.3
msedge-selenium-tools==3.141.3
Pillow==8.2.0
pycryptodome==3.10.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
requests==2.25.1
selenium==3.141.0
six==1.15.0
urllib3==1.26.4
wincertstore==0.2
ffmpeg-python==0.2.0
